analytic:chore - Update lint e make commands
I'll open a pull request for each service updating its
make commands and already updating the project's lint.
The basis of the structure is to update all services to
use golang in version 1.17 and get `.golangci.yml` file
from horusec-devkit project and usage it here.
In this commit I updated the service `analytic`.

Signed-off-by: wilian <wilian.silva@zup.com.br>
Signed-off-by: Wilian Gabriel <wilian.silva@zup.com.br>